# laravel/source/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/productsale": {
            "get": {
                "tags": [
                    "ProductSale"
                ],
                "summary": "Get All ProductSale",
                "operationId": "e13d73239ccb90b63a1c0612ffb3c850",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "ProductSale"
                ],
                "summary": "Add ProductSale",
                "operationId": "fb0ae373fd8a6e89e23931b3cac36995",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "productid": {
                                        "type": "integer"
                                    },
                                    "saleid": {
                                        "type": "integer"
                                    },
                                    "deleted": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/productsale/{id}": {
            "get": {
                "tags": [
                    "ProductSale"
                ],
                "summary": "Get ProductSale",
                "operationId": "e85bc0fe2d202fd3c063a4e9b4e00ee9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "ProductSale"
                ],
                "summary": "Update ProductSale",
                "operationId": "2a3e11181e6a8ba4b6f93268fc492080",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "productid": {
                                        "type": "integer"
                                    },
                                    "saleid": {
                                        "type": "integer"
                                    },
                                    "deleted": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "ProductSale"
                ],
                "summary": "Delete ProductSale",
                "operationId": "1d8ed2c591bbebd5ecad3d1bc74a6259",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/receipt": {
            "get": {
                "tags": [
                    "Receipt"
                ],
                "summary": "Get All Receipt",
                "operationId": "7a58e266cdd0ac1bcc799958c4d03c51",
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "post": {
                "tags": [
                    "Receipt"
                ],
                "summary": "Add Receipt",
                "operationId": "ba87ba0047e1b80c64588f5b9446f820",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "string"
                                    },
                                    "timeOrder": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "paymentTime": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "paymentMethod": {
                                        "type": "string"
                                    },
                                    "deleted": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
        "/api/receipt/{id}": {
            "get": {
                "tags": [
                    "Receipt"
                ],
                "summary": "Get Receipt",
                "operationId": "c2c4109534951f8ea9ff15f6fe7c99a5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "put": {
                "tags": [
                    "Receipt"
                ],
                "summary": "Update Receipt",
                "operationId": "1f5e86903246fe63c9a5fa1d37d04f33",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    },
                                    "user_id": {
                                        "type": "string"
                                    },
                                    "timeOrder": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "paymentTime": {
                                        "type": "string",
                                        "format": "date-time"
                                    },
                                    "paymentMethod": {
                                        "type": "string"
                                    },
                                    "deleted": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Receipt"
                ],
                "summary": "Delete Receipt",
                "operationId": "ebe0ea885826f280cfb46cc6f7cba0e0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": false,
                        "schema": {
                            "type": "int"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        },
...
    }
}